Bug Fixes, Add user exceptions, Add main body.
Added bug fixes for SQL queries. Added exceptions table in update_user_by_username function.
Added main body, the file now queries t.me for every user to check the integrity of the chain. Then finds the longest chain, converts it to string and compares it to last sent chain. If the chains aren't same, it sends it and saves it in its place.